requirements.txt
torch~=2.6.0
torchvision~=0.21.0
retinaface-pytorch~=0.0.8
torchao~=0.9.0
torchinfo~=1.8.0